Dividing the main file into individual files
So there isn't a million changes to  the main.c file, I divided up the previous main file into 3 files. The setup file contains the setup subroutines involving timers, IO ports, ect. The animations file contains what the LEDs will do. This will allow anyone to track individual changes more easily. There might be provisions for a future fill that handles inputs, but that might just be divided into setup and main.
