+++ setup.h
@@ -1,0 +1,32 @@
+/* 
+ * File:   setup.h
+ * Author: church
+ *
+ * Created on April 5, 2019, 5:24 PM
+ */
+
+#ifndef SETUP_H
+#define	SETUP_H
+
+#ifdef	__cplusplus
+extern "C" {
+#endif
+
+
+
+
+#ifdef	__cplusplus
+}
+#endif
+//global variables
+
+//functions
+void LED_Setup(void);
+void Timer_and_Interrupt_setup (void);
+void setup(void);
+
+
+
+
+#endif	/* SETUP_H */
+
